AAAD-37 Refine file structure
