+++ src/pages/ProfilePage/Profile.tsx
@@ -1,5 +1,5 @@
 import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar } from '@ionic/react';
-import ExploreContainer from '../components/ExploreContainer';
+import ExploreContainer from '../../components/ExploreContainer';
 import './Profile.css';
 
 const ProfileView: React.FC = () => {
